src/requirements.txt
Flask==2.3.3
pymongo==4.5.0
python-dotenv==1.0.0
flask-login==0.6.2
werkzeug==2.3.7
email-validator==2.0.0
flask-wtf==1.1.1
wtforms==3.0.1
dnspython==2.4.2
python-dateutil==2.8.2